refactor with JPA

+++ homework-6/src/main/java/ru/otus/homework/service/IOService.java
@@ -5,5 +5,6 @@ public interface IOService {
     String read();
     Integer readInt();
 
-    void write(int count);
+    void write(long count);
+
 }
